Add threads
The -p option controls how many threads we want to use for mining. Each
thread will be given a sequence of nonces to mine that is disjoint from
all other threads. Each thread reports independently on the best nonces
it has found, and we pick any of the best ones at the end.